test-requirements.txt
django-fake-model==0.1.4
django-querycount==0.5.0
freezegun==0.3.9
httpretty==0.8.14
mock==2.0.0
pytest-django==2.9.1
pytest-mock==0.10.1
pytest-pythonpath==0.7.1
pytest==2.8.7
pytz==2015.7
PyYAML==3.11
